Add files via upload
Goal
The goal of this assignment is to practice with code analysis tools. You will analyze the potential vulnerabilities of a set of simple programs using two publicly-available code analysis tools on a set of programs and assess the validity of the findings.

Environment setup

1- Install code analysis tool Cppcheck. The tool is available at: http://cppcheck.sourceforge.netLinks to an external site.

2- Install code analysis tool Clang. The tool is available at: http://clang.orgLinks to an external site.

 

Exercise 1 
Archive CSCE5565-Assignment2Files.zip Download CSCE5565-Assignment2Files.zipincludes 5 C programs.  Your task is to use CppCheck and Clang to analyze the potential security vulnerabilities of these programs. The steps are:

1- Run CppCheck on each of the 5 programs and report the outputs of the analysis of each of the programs. The command line to run CppCheck is: cppcheck --enable=all --inconclusive --library=posix <file name>. Enumerate the potential security vulnerabilities in each of the programs, if any.

2- Run Clang of each of the 5 programs and report the outputs of the analysis of each of the programs. The command line to run Clang is:  clang -O -g -fsanitize=address <file name> && ./a.out. Enumerate the potential security vulnerabilities in each of the programs, if any.

3- Compare the results obtained by both tools for each program. Review manually the code of each program to assess the validity (true positive, false positive, false negative) of each of the findings returned by the tools.  

Exercise 2
The National Institute of Standards and Technology has developed a database of security test cases that organizations can use to assess the performance of code analysis tools. The database is called NIST Software Assurance Reference Dataset Project and is available at https://samate.nist.gov/SARD/testsuite.phpLinks to an external site.. 

Review the website and select two security test case of your choice from SARD Suites/Juliet 1.3.1 C/C++ test suite supportLinks to an external site.. Use the test case to assess CppCheck and Clang. Report the findings of the tools and your interpretation of the validity of the findings.

Note: You need to write a program that includes the test case as a function and a main() in order to use the tools--the tools analyze programs and not functions.